refactor(tools): types refactor
