+++ src/ethereum_test_tools/__init__.py
@@ -75,12 +75,9 @@ __all__ = (
     "Conditional",
     "EngineAPIError",
     "Environment",
-    "Fixture",
     "FixtureCollector",
-    "FixtureEngineNewPayload",
     "Header",
     "HistoryStorageAddress",
-    "HiveFixture",
     "Initcode",
     "JSONEncoder",
     "Opcode",
