+++ src/ethereum_test_tools/common/types.py
@@ -856,39 +856,6 @@ def withdrawals_root(withdrawals: List[Withdrawal]) -> bytes:
     return t.root_hash
 
 
-@dataclass(kw_only=True)
-class FixtureWithdrawal(Withdrawal):
-    """
-    Structure to represent a single withdrawal of a validator's balance from
-    the beacon chain in the output fixture.
-    """
-
-    index: NumberConvertible = field(
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    validator: NumberConvertible = field(
-        json_encoder=JSONEncoder.Field(
-            name="validatorIndex",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    amount: NumberConvertible = field(
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-
-    @classmethod
-    def from_withdrawal(cls, w: Withdrawal) -> "FixtureWithdrawal":
-        """
-        Returns a FixtureWithdrawal from a Withdrawal.
-        """
-        kwargs = {field.name: getattr(w, field.name) for field in fields(w)}
-        return cls(**kwargs)
-
-
 DEFAULT_BASE_FEE = 7
 
 
@@ -1772,114 +1739,3 @@ def blob_versioned_hashes_from_transactions(
             versioned_hashes.extend(tx.blob_versioned_hashes)
 
     return versioned_hashes
-
-
-@dataclass
-class FixtureTransaction(Transaction):
-    """
-    Representation of an Ethereum transaction within a test Fixture.
-    """
-
-    ty: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            name="type",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    """
-    Transaction type value.
-    """
-    chain_id: int = field(
-        default=1,
-        json_encoder=JSONEncoder.Field(
-            name="chainId",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    nonce: int = field(
-        default=0,
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    gas_price: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            name="gasPrice",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    max_priority_fee_per_gas: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            name="maxPriorityFeePerGas",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    max_fee_per_gas: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            name="maxFeePerGas",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    gas_limit: int = field(
-        default=21000,
-        json_encoder=JSONEncoder.Field(
-            name="gasLimit",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    to: Optional[FixedSizeBytesConvertible] = field(
-        default=AddrAA,
-        json_encoder=JSONEncoder.Field(
-            cast_type=Address,
-            default_value_skip_cast="",
-        ),
-    )
-    value: int = field(
-        default=0,
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    data: BytesConvertible = field(
-        default_factory=bytes,
-        json_encoder=JSONEncoder.Field(
-            cast_type=Bytes,
-        ),
-    )
-    max_fee_per_blob_gas: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            name="maxFeePerBlobGas",
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    v: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    r: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-    s: Optional[int] = field(
-        default=None,
-        json_encoder=JSONEncoder.Field(
-            cast_type=ZeroPaddedHexNumber,
-        ),
-    )
-
-    @classmethod
-    def from_transaction(cls, tx: Transaction) -> "FixtureTransaction":
-        """
-        Returns a FixtureTransaction from a Transaction.
-        """
-        kwargs = {field.name: getattr(tx, field.name) for field in fields(tx)}
-        return cls(**kwargs)
